+++ sdk/nodejs/keyvault/secret.ts
@@ -46,9 +46,7 @@ import * as utilities from "../utilities";
  *     location: testResourceGroup.location,
  *     name: server.hex.apply(hex => sprintf.sprintf("%s%s", "kv", hex)),
  *     resourceGroupName: testResourceGroup.name,
- *     sku: {
- *         name: "premium",
- *     },
+ *     skuName: "premium",
  *     tags: {
  *         environment: "Production",
  *     },